+++ src/BaseCommand.ts
@@ -1,14 +1,14 @@
-import Command from '@oclif/command'
-import chalk from 'chalk'
+import Command from "@oclif/command";
+import chalk from "chalk";
 
 export default abstract class extends Command {
-  
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
   async catch(err: Error) {
-    if(err.message) {
-      return console.error(chalk.red(err.message));
+    if (err.message) {
+      console.error(chalk.red(err.message));
     }
-    if(typeof err === "string") {
-      return console.error(chalk.red(err));
+    if (typeof err === "string") {
+      console.error(chalk.red(err));
     }
 
     // console.error(chalk.red(JSON.stringify(err)));