+++ src/commands/init.ts
@@ -211,18 +211,18 @@ export class InitCommand extends Command {
         message: "Select your interpolation",
         type: "list",
         choices: [
-          { id: "polyglot", name: "%{ variable }" },
-          { id: "i18next", name: "{{ variable }}" },
-          { id: "i18n", name: "{ variable }" },
+          { value: "polyglot", name: "%{ variable }" },
+          { value: "i18next", name: "{{ variable }}" },
+          { value: "i18n", name: "{ variable }" },
           {
-            id: "icu",
+            value: "icu",
             name: "ICU Messages, i.e. { variable, number, ::currency/USD }",
           },
           // eslint-disable-next-line
-          { id: "ruby", name: "${ variable }" },
-          { id: "laravel", name: ":variable" },
-          { id: "default", name: "Automatic" },
-          { id: "none", name: "None" },
+          { value: "ruby", name: "${ variable }" },
+          { value: "laravel", name: ":variable" },
+          { value: "default", name: "Automatic" },
+          { value: "none", name: "None" },
         ],
       },
       {