+++ CHANGELOG.md
@@ -6,6 +6,14 @@ Simpleen is a [translation platform for i18n projects](https://simpleen.io/).
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2021-02-17
+
+- Add new dataformats to translate .yml,.yaml,.po and .json
+- Fix error in init in config - results in invalid interpolations
+- Ignore lock file in translate command
+- Add $PATH variable in output_path
+- Upgrade dependencies
+
 ## [0.3.0] - 2021-02-01
 
 - Add new languages supported by google translate